requirements.txt
Django==5.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
drf-spectacular==0.27.0
markdown==3.5.1
django-filter==23.5
python-dotenv==1.0.0
pillow==10.1.0
django-taggit==4.0.0
django-ckeditor==6.7.1  # Updated to latest version with CKEditor 4.25.0-lts
django-cors-headers==4.3.1
